refactor: response 폴더 및 파일 위치 변경
공통은 constants 하위로, domain에 해당하는 파일들은 각 domain 하위로 이동
