+++ server/src/constants/response/common.ts
@@ -1,4 +1,4 @@
-import { failRes } from '@response/index';
+import { failRes } from './index';
 
 export const COMMON_EXCEPTION = {
   INVALID_QUERY_PARAMS: failRes('올바르지 않은 Query Params입니다.'),
