+++ server/src/constants/response/location.ts
