+++ server/src/map/map.response.ts
